Model Deployment/requirements.txt
Flask==2.0.1
h5py==3.1.0
Keras-Preprocessing==1.1.2
numpy==1.19.5
Pillow==8.3.1
requests==2.22.0
requests-oauthlib==1.3.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
urllib3==1.25.7
Werkzeug==2.0.1
